refactor(packages): Split pandocast inputter and package
